feat: add search drawer, nav and footer comps
